+++ tests/integration.rs
@@ -10,8 +10,8 @@ fn resonator(seqs: &[&str], max_dist: u32) -> HammingResonator {
     HammingResonator::with_max_dist(v, max_dist).unwrap()
 }
 
-fn weighted(seqs: &[(&str, f64)], max_dist: u32) -> HammingResonatorWeighted {
-    let v: Vec<(BString, f64)> = seqs.iter().map(|&(s, w)| (BString::from(s), w)).collect();
+fn weighted(seqs: &[(&str, f32)], max_dist: u32) -> HammingResonatorWeighted {
+    let v: Vec<(BString, f32)> = seqs.iter().map(|&(s, w)| (BString::from(s), w)).collect();
     HammingResonatorWeighted::with_max_dist(v, max_dist).unwrap()
 }
